scripts/requirements.txt
litellm
tabulate
termcolor
python-dotenv